+++ packages/dataprotector-consumer/jest.config.cjs
@@ -1,0 +1,13 @@
+/** @type {import('ts-jest').JestConfigWithTsJest} */
+module.exports = {
+  testEnvironment: 'node',
+  extensionsToTreatAsEsm: ['.ts'],
+  testTimeout: 10000,
+  moduleNameMapper: {
+    '^(\\.{1,2}/.*)\\.js$': '$1',
+  },
+  transform: {
+    // https://swc.rs/docs/usage/jest
+    '^.+\\.(t|j)sx?$': '@swc/jest',
+  },
+};