+++ packages/sdk/src/lib/types/dataProtectorTypes.ts
@@ -12,6 +12,7 @@ import {
 export type DataScalarType =
   | boolean
   | number
+  | bigint
   | string
   | Uint8Array
   | ArrayBuffer;