+++ packages/subgraph/src/dataProtector.ts
@@ -11,9 +11,12 @@ import { AUTHORIZED_CHARACTERS } from './types';
 
 const PATH_SEPARATOR = '.';
 const DataSchemaEntryType = [
-  'boolean',
-  'number',
+  'boolean', // dataprotector v0 schema for boolean 
+  'number', // dataprotector v0 schema for integers in safe integers range
   'string',
+  'bool',
+  'i128',
+  'f64',
   // MIME types supported by dataprotector
   'application/octet-stream',
   'application/pdf',